refactor(cli,review,session): fix three code-review findings from perf-1-file-size
STYLE-1: break circular static import between plan.ts and plan-decompose.ts
- extract DEFAULT_TIMEOUT_SECONDS, createPlanRuntime, resolvePlanModelSelection,
  and _planDeps into plan-runtime.ts
- plan-decompose.ts now imports from ./plan-runtime (no longer from ./plan)
- plan.ts re-exports the moved symbols for backward compatibility so existing
  callers and tests continue to work without changes

STYLE-2: pre-emptively split manager.ts (599/600 lines)
- extract sweepOrphans body + DEFAULT_ORPHAN_TTL_MS into manager-sweep.ts
- manager.ts delegates in a single line: sweepOrphansImpl(this._sessions, ttlMs)
- gives manager.ts ~30 lines of headroom below the 600-line limit

ENH-1: deduplicate SEVERITY_RANK across review helpers
- introduce src/review/severity.ts as the single source of truth
- semantic-helpers.ts and adversarial-helpers.ts import from ./severity
  instead of each defining an identical local constant
